@@ -12,17 +12,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "2.0.0",
-    "@angular/compiler": "2.0.0",
-    "@angular/core": "2.0.0",
-    "@angular/forms": "2.0.0",
-    "@angular/http": "2.0.0",
-    "@angular/platform-browser": "2.0.0",
-    "@angular/platform-browser-dynamic": "2.0.0",
-    "@angular/router": "3.0.0",
+    "@angular/common": "^2.1.2",
+    "@angular/compiler": "^2.1.2",
+    "@angular/core": "^2.1.2",
+    "@angular/http": "^2.1.2",
+    "@angular/platform-browser": "^2.1.2",
+    "@angular/platform-browser-dynamic": "^2.1.2",
+    "@angular/compiler-cli": "^2.1.2",
+    "@angular/router": "^3.1.2",
     "bootstrap": "^4.0.0-alpha.5",
     "core-js": "^2.4.1",
     "ng2-redux": "^4.1.0",
+    "ng2-redux-router": "^2.0.0",
     "redux": "^3.6.0",
     "redux-logger": "^2.7.4",
     "rxjs": "5.0.0-beta.12",
