@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { InternshipsService } from './internships.service';
 import { FilterInternships } from './search-internship.filter';
 import { Router } from '@angular/router';
@@ -17,65 +17,23 @@ import { InternshipState, IAppState } from './store';
     }
     `],
     template: `
-    <div class="row" style="margin-bottom: 20px;">
-        <div class="col-xs-6">
-            <button routerLink="/internship" class="btn btn-primary">Create new internship</button>
+    <div class="col-sm-6 col-xl-4" *ngFor="let internship of internships">
+        <div *ngIf="internship === -1">
+            "No matches"
         </div>
-        <div class="float-xs-right col-xs-6 col-sm-5 col-md-4 col-lg-3 col-xl-2">
-          <input class="form-control" type="text" placeholder="Search" [(ngModel)]="search">
-        </div>
-    </div>
-
-    <div *ngIf="message$ | async">
-        Retrieving data. Please wait..
-    </div>
-
-    <div class="row">
-        <div class="col-sm-6 col-xl-4" *ngFor="let internship of (internships$ | async).internships">
-
-            <div *ngIf="internship === -1">
-                "No matches"
-            </div>
-            <div *ngIf="internship !== -1" class="card card-block" >
-                <h3 class="card-title">{{internship.student.firstname}} {{internship.student.lastname}}</h3>
-                <p class="card-text">{{internship.companyName}}</p>
-                <a (click)="gotoDetails(internship._id)" class="btn btn-primary">Details</a>
-            </div>
-        </div>
-    </div>    `
-})
-export class InternshipsComponent implements OnInit {
-    // private internships: Observable<InternshipState>;
-    private message$: Observable<Boolean>;
-    
-    @select('internships') internships$: Observable<InternshipState>;
-
-
-    constructor(private internshipsService: InternshipsService, public actions: InternshipActions,
-        private router: Router, private ngRedux: NgRedux<IAppState>) {
-    }
-    
-    ngOnInit():void {
-        // console.log("component init");
-        this.actions.getInternships();
-        this.message$ = this.ngRedux.select(state => state.internships.isFetching); //selecting a specific part of the state
-
-        // this.internships = this.internshipsService.getAllLocalInternships();
-        
-        // if (!this.internships) {
-        //     this.message = "Retrieving data...";
+        <div *ngIf="internship !== -1" class="card card-block" >
+            <h3 class="card-title">{{internship.student?.firstname}} {{internship.student?.lastname}}</h3>
+            <p class="card-text">{{internship.companyName}}</p>
             
-        //     this.internshipsService.getAllInternships().subscribe(
-        //     (internships) => {
-        //         this.internships = internships;
-        //         this.message = "";
-        //     },
-        //         error => this.message = error 
-        //     );
-        // }
-    }
+            <a (click)="gotoDetails.emit(internship._id)" class="btn btn-primary">Details</a>
+        </div>
+    </div>
+    `
+})
+export class InternshipsComponent {
+    @Input() internships: Internship[];    
+    @Output() gotoDetails: EventEmitter<any> = new EventEmitter();
 
-    gotoDetails(id: string) {
-        this.router.navigate(['/internship', id]);
+    constructor() {
     }
 } 
